pyproject.toml
[tool.poetry]
name = "llm-mushroom-chatbot"
version = "0.1.0"
description = ""
authors = ["Tage Johansson <tajo9185@student.uu.se>"]
readme = "readme.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
google-generativeai = "^0.8.1"
gradio = "^4.44.0"
ipywidgets = "^8.1.5"
pillow = "^10.4.0"


[tool.poetry.group.dev.dependencies]
notebook = "^7.2.2"
pyright = "^1.1.381"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
